# eyeson-go-server/static/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "EyesOn SIM Management API",
...
    "version": "1.0.0",
    "contact": {
      "name": "EyesOn API Support",
      "email": "api-support@eyeson.local",
      "url": "https://github.com/alexgavs/eyeson-go"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "x-logo": {
      "url": "https://example.com/logo.png",
      "altText": "EyesOn API"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local Development Server"
    },
    {
      "url": "http://127.0.0.1:5000",
      "description": "Local Server (IP Address)"
    },
    {
      "url": "https://api.eyeson.local",
      "description": "Production Server (Example)"
    }
  ],
...
  "paths": {
...
    "/api/v1/queue/my": {
      "get": {
        "tags": ["Queue"],
        "summary": "Get current user's active tasks",
        "description": "Returns a list of the authenticated user's pending and processing tasks in the queue.",
        "operationId": "getMyQueue",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "User's active tasks retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueTaskList"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/queue/task/{id}": {
      "get": {
        "tags": ["Queue"],
        "summary": "Get task status by ID",
        "description": "Returns detailed information about a specific task in the queue.",
        "operationId": "getTaskStatus",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Task ID",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task details retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueTask"
                }
              }
            }
          },
          "404": {
            "description": "Task not found"
          }
        }
      }
    },
    "/api/v1/queue/task/{id}/cancel": {
      "post": {
        "tags": ["Queue"],
        "summary": "Cancel own task",
        "description": "Cancels a pending or processing task owned by the current user.",
        "operationId": "cancelTask",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Task ID to cancel",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task cancelled successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Task cannot be cancelled"
          },
          "404": {
            "description": "Task not found or not owned by user"
          }
        }
      }
    },
    "/api/v1/queue/stats": {
      "get": {
        "tags": ["Queue"],
        "summary": "Get queue statistics (Admin)",
        "description": "Returns overall queue statistics including pending, processing, completed, and failed task counts.\n\n**Required Role**: Administrator",
        "operationId": "getQueueStats",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Queue statistics retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueStats"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Administrator role required"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JSON Web Token authentication. Obtain a token via POST /api/v1/auth/login and include it in the Authorization header."
      }
    },
    "schemas": {
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Success message",
            "example": "Operation completed successfully"
          }
        }
      },
...
      "QueueTask": {
        "type": "object",
        "description": "Task in the processing queue",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique task identifier",
            "example": 123
          },
          "type": {
            "type": "string",
            "enum": ["LABEL_UPDATE", "STATUS_CHANGE"],
            "description": "Type of operation"
          },
          "status": {
            "type": "string",
            "enum": ["PENDING", "PROCESSING", "COMPLETED", "FAILED", "CANCELLED"],
            "description": "Current task status"
          },
          "priority": {
            "type": "string",
            "enum": ["LOW", "NORMAL", "HIGH", "CRITICAL"],
            "description": "Task priority level"
          },
          "target_msisdn": {
            "type": "string",
            "description": "Target SIM card MSISDN",
            "example": "972501234567"
          },
          "target_cli": {
            "type": "string",
            "description": "Target SIM card CLI"
          },
          "label_field": {
            "type": "string",
            "description": "Label field being updated (for LABEL_UPDATE tasks)",
            "example": "label_1"
          },
          "label_value": {
            "type": "string",
            "description": "New label value"
          },
          "old_status": {
            "type": "string",
            "description": "Previous status (for STATUS_CHANGE tasks)"
          },
          "new_status": {
            "type": "string",
            "description": "Target status (for STATUS_CHANGE tasks)"
          },
          "attempt": {
            "type": "integer",
            "description": "Current attempt number",
            "example": 1
          },
          "max_attempts": {
            "type": "integer",
            "description": "Maximum retry attempts",
            "example": 3
          },
          "last_error": {
            "type": "string",
            "description": "Last error message if failed"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Task creation timestamp"
          },
          "completed_at": {
            "type": "string",
            "format": "date-time",
            "description": "Task completion timestamp"
          },
          "username": {
            "type": "string",
            "description": "User who created the task"
          }
        }
      },
      "QueueTaskList": {
        "type": "object",
        "properties": {
          "tasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QueueTask"
            }
          },
          "total": {
            "type": "integer",
            "description": "Total number of tasks"
          }
        }
      },
      "QueueStats": {
        "type": "object",
        "description": "Queue statistics",
        "properties": {
          "pending": {
            "type": "integer",
            "description": "Tasks waiting to be processed",
            "example": 5
          },
          "processing": {
            "type": "integer",
            "description": "Tasks currently being processed",
            "example": 1
          },
          "completed": {
            "type": "integer",
            "description": "Successfully completed tasks",
            "example": 150
          },
          "failed": {
            "type": "integer",
            "description": "Failed tasks",
            "example": 3
          },
          "cancelled": {
            "type": "integer",
            "description": "Cancelled tasks",
            "example": 2
          },
          "today_total": {
            "type": "integer",
            "description": "Total tasks created today",
            "example": 25
          }
        }
      }
    }
  }
}